split types into types file
