+++ x-pack/solutions/observability/plugins/slo/public/pages/slo_details/components/definition/types.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { SLOWithSummaryResponse } from '@kbn/slo-schema';
+
+export interface SloDetailsDefinitionProps {
+  slo: SLOWithSummaryResponse;
+}
